+++ packages/sdk/src/lib/plugins.ts
@@ -1,8 +1,9 @@
 import deno from '../plugins/deno.ts'
+import githubActions from '../plugins/github-actions.ts'
 import npm from '../plugins/npm.ts'
 import pnpm from '../plugins/pnpm.ts'
 import ruby from '../plugins/ruby.ts'
 import uv from '../plugins/uv.ts'
 import yarn from '../plugins/yarn.ts'
 
-export default { deno, npm, pnpm, ruby, uv, yarn }
+export default { deno, githubActions, npm, pnpm, ruby, uv, yarn }